Rename Collection to Namespace
Move Namespace to resource.py
