+++ replicate/deployment.py
@@ -1,10 +1,9 @@
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
 
-from replicate.collection import Collection
 from replicate.files import upload_file
 from replicate.json import encode_json
 from replicate.prediction import Prediction
-from replicate.resource import Resource
+from replicate.resource import Namespace, Resource
 
 if TYPE_CHECKING:
     from replicate.client import Client
@@ -15,7 +14,7 @@ class Deployment(Resource):
     A deployment of a model hosted on Replicate.
     """
 
-    _collection: "DeploymentCollection"
+    _namespace: "DeploymentCollection"
 
     username: str
     """
@@ -36,7 +35,7 @@ class Deployment(Resource):
         return DeploymentPredictionCollection(client=self._client, deployment=self)
 
 
-class DeploymentCollection(Collection):
+class DeploymentCollection(Namespace):
     """
     Namespace for operations related to deployments.
     """
@@ -66,7 +65,7 @@ class DeploymentCollection(Collection):
         return super()._prepare_model(attrs)
 
 
-class DeploymentPredictionCollection(Collection):
+class DeploymentPredictionCollection(Namespace):
     """
     Namespace for operations related to predictions in a deployment.
     """
