+++ replicate/training.py
@@ -3,11 +3,10 @@ from typing import Any, Dict, List, Optional, TypedDict, Union
 
 from typing_extensions import NotRequired, Unpack, overload
 
-from replicate.collection import Collection
 from replicate.exceptions import ReplicateException
 from replicate.files import upload_file
 from replicate.json import encode_json
-from replicate.resource import Resource
+from replicate.resource import Namespace, Resource
 from replicate.version import Version
 
 
@@ -16,7 +15,7 @@ class Training(Resource):
     A training made for a model hosted on Replicate.
     """
 
-    _collection: "TrainingCollection"
+    _namespace: "TrainingCollection"
 
     id: str
     """The unique ID of the training."""
@@ -69,12 +68,12 @@ class Training(Resource):
         Load the training from the server.
         """
 
-        obj = self._collection.get(self.id)  # pylint: disable=no-member
+        obj = self._namespace.get(self.id)  # pylint: disable=no-member
         for name, value in obj.dict().items():
             setattr(self, name, value)
 
 
-class TrainingCollection(Collection):
+class TrainingCollection(Namespace):
     """
     Namespace for operations related to trainings.
     """
